Game Directory Function
- Added support to link the game directory with the add-on, and auto populate to the Giants Exporter Tool for Blender.
- Renamed to "FS25 Modding Extension" to better describe what it does.
